server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.7.0
pypdf==4.2.0
python-docx==1.1.2
sentence-transformers==2.7.0
numpy==1.26.4
groq==0.8.0
pillow==10.4.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84